+++ postgres-protocol/src/types/test.rs
@@ -6,6 +6,7 @@ use super::*;
 use crate::IsNull;
 
 #[test]
+#[allow(clippy::bool_assert_comparison)]
 fn bool() {
     let mut buf = BytesMut::new();
     bool_to_sql(true, &mut buf);
@@ -113,7 +114,7 @@ fn array() {
     .unwrap();
 
     let array = array_from_sql(&buf).unwrap();
-    assert_eq!(array.has_nulls(), true);
+    assert!(array.has_nulls());
     assert_eq!(array.element_type(), 10);
     assert_eq!(array.dimensions().collect::<Vec<_>>().unwrap(), dimensions);
     assert_eq!(array.values().collect::<Vec<_>>().unwrap(), values);
@@ -150,7 +151,7 @@ fn non_null_array() {
     .unwrap();
 
     let array = array_from_sql(&buf).unwrap();
-    assert_eq!(array.has_nulls(), false);
+    assert!(!array.has_nulls());
     assert_eq!(array.element_type(), 10);
     assert_eq!(array.dimensions().collect::<Vec<_>>().unwrap(), dimensions);
     assert_eq!(array.values().collect::<Vec<_>>().unwrap(), values);