+++ tokio-postgres/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Change Log
 
+## v0.7.5 - 2021-10-29
+
+### Fixed
+
+* Fixed a bug where the client could enter into a transaction if the `Client::transaction` future was dropped before completion.
+
+## v0.7.4 - 2021-10-19
+
+### Fixed
+
+* Fixed reporting of commit-time errors triggered by deferred constraints.
+
+## v0.7.3 - 2021-09-29
+
+### Fixed
+
+* Fixed a deadlock when pipelined requests concurrently prepare cached typeinfo queries.
+
+### Added
+
+* Added `SimpleQueryRow::columns`.
+* Added support for `eui48` 1.0 via the `with-eui48-1` feature.
+* Added `FromSql` and `ToSql` implementations for arrays via the `array-impls` feature.
+* Added support for `time` 0.3 via the `with-time-0_3` feature.
+
 ## v0.7.2 - 2021-04-25
 
 ### Fixed